Add suport for long and double

+++ src/main/java/se/lindhen/qrgame/bytecode/FriendlyWriter.java
@@ -11,7 +11,7 @@ public class FriendlyWriter extends BitWriter {
     }
 
     @Override
-    public void write(int size, int value) {
+    public void write(int size, long value) {
         stringBuffer.append("(");
         stringBuffer.append(size);
         stringBuffer.append(":");
